Moved all tasks (simulation objects) from the emulator to the SDK
