+++ lightworks/__init__.py
@@ -48,6 +48,7 @@ from .__settings import settings
 from .__version import __version__
 from .sdk.circuit import Circuit, Parameter, ParameterDict, Unitary
 from .sdk.state import State
+from .sdk.tasks import Analyzer, QuickSampler, Sampler, Simulator
 from .sdk.utils import (
     PostSelection,
     PostSelectionFunction,
@@ -67,9 +68,10 @@ except ModuleNotFoundError:
 
 # fmt: off
 __all__ = [
-    "emulator", "qubit", "interferometers", "tomography",
+    "emulator", "qubit", "interferometers", "tomography", "settings",
     "Circuit", "Unitary", "Display", "State", "random_unitary",
     "random_permutation", "db_loss_to_decimal", "decimal_to_db_loss",
-    "Parameter", "ParameterDict", "PostSelection", "PostSelectionFunction"
+    "Parameter", "ParameterDict", "PostSelection", "PostSelectionFunction",
+    "Simulator", "Sampler", "Analyzer", "QuickSampler"
 ]
 # fmt: on
