+++ lightworks/emulator/__init__.py
@@ -46,12 +46,6 @@ Simulators:
 
 from .backends import Backend
 from .components import Detector, Source
-from .simulation import Analyzer, QuickSampler, Sampler, Simulator
 from .utils.exceptions import *
 
-# fmt: off
-__all__ = [
-    "Simulator", "Sampler", "Source", "Detector", "Analyzer", "QuickSampler",
-    "Backend"
-]
-# fmt: on
+__all__ = ["Source", "Detector", "Backend"]
