+++ lightworks/emulator/utils/__init__.py
@@ -15,5 +15,4 @@
 # ruff: noqa: F401, F403
 
 from .exceptions import *
-from .post_selection_processing import process_post_selection
-from .state_utils import annotated_state_to_string, fock_basis
+from .state_utils import annotated_state_to_string
