+++ lightworks/sdk/results/simulation_result.py
@@ -19,7 +19,7 @@ import matplotlib.pyplot as plt
 import numpy as np
 import pandas as pd
 
-from ...sdk.state import State
+from ..state import State
 from ..utils import ResultCreationError
