+++ lightworks/sdk/tasks/simulator.py
@@ -15,12 +15,16 @@
 
 import numpy as np
 
-from ...sdk.circuit import Circuit
-from ...sdk.state import State
-from ...sdk.utils import add_heralds_to_state
-from ..backends.fock_backend import FockBackend
+from ...emulator.backends.fock_backend import FockBackend
+from ..circuit import Circuit
 from ..results import SimulationResult
-from ..utils import ModeMismatchError, PhotonNumberError, fock_basis
+from ..state import State
+from ..utils import (
+    ModeMismatchError,
+    PhotonNumberError,
+    add_heralds_to_state,
+    fock_basis,
+)
 from .task import Task
